refactor: move UI events to separate file
